Fix MX-record update bug.
The RDATA needs to be compared, so DNSRecord.d_content has a nice getZoneRepresentation()
but the DNSResourceRecord does not. Added that to it so we can compare.
With MX records, selecting based on priority was also needed.
We don't have a situation where the priority is updated, as it's part of the RDATA.