Use the YYYYMMDDXX format for the serial number.
When we need to update the serial number, we compare the
current serial with YYYYMMDD01. If the current serial is newer, we simply
increase that by one. If it's older, we set the serial to YYYYMMDD01.

When using SOA-EDIT you might still get an old serial.
The advice is then to run with INCEPTION-EPOCH, as that updates the serial only if it's newer.