Fix in-between and delegate test
The purging of the cache was broken. This was due to a wrong test
(see one of the previous commits).
When that was fixed, the bind-dnssec test failed because it only
refreshes the zone on 'commit' - resulting in getBeforeAndAfter()
returning the wrong content to purge the cache.
This is now resolved by giving the getBeforeAndAfter() an option
to really get the before (and not the current). All 'supported rfc2136'
backends have been updated for this.

After this, the delegate test was broken because for delegate a remove
might mean that there is now a new before/after. So we get all the befores
and afters prior to when we do all the updates. We then purge the cache
with those before/afters. Finally resulting in the correct behaviour.