Add rfc2136 implementation with replaceRRSEt and dirty cache clean
This completes the RFC2136 implementation and passes all the dyndns tests using the replaceRRSet method.

The runtest now understands a 'once' option because a SOA-update can only be performed once.
dnsbackend.hh has a listSubZone() - this is not a pretty method, but it's required to fix delegation flags without
reading the complete zone.
replaceRRSet is also updated to understand 'ANY' as a Qtype. This performs the query without the where type-cause. It also allows for not adding any items which makes it usable to delete RRSET's.