Skip all dyndns tests when testing presigned
The test are performed on the slave, which forwards the package to the master.
This in theorie should make it work, but for the slave to receive the update, means we have to wait for the slave to receive an update of the domain. This is something that impacts the testset a bit too much and thus all dyndns tests are skipped when running in presigned more.